% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSorters.R
\name{printSorters}
\alias{printSorters}
\title{Print and Return Unique Sorters in Concept Map Data}
\usage{
printSorters(CMData, verbose = TRUE)
}
\arguments{
\item{CMData}{A data frame containing concept map data. This must include a column named \code{"sorterID"}.}

\item{verbose}{A logical, if \code{TRUE}, the function will print the list of unique sorters to the console.}
}
\value{
A vector of unique sorter IDs.
}
\description{
This function retrieves and optionally prints the unique sorters (users) in a given concept mapping dataset.
}
\details{
The function first checks if the provided dataset is suitable for concept mapping using the \cr
\code{checkConceptMapData} function. If the data is valid, it retrieves the unique sorter IDs
from the \code{sorterID} column. If \code{verbose = TRUE}, the function prints the sorter IDs.
}
\examples{
# Example of valid data
CMData <- data.frame(
    sorterID = c("resp1", "resp1", "resp1", "resp2",
     "resp2", "resp2", "resp3", "resp3", "resp3"),
   statement = c("London", "Frankfurt", "Berlin", "London",
    "Frankfurt", "Berlin", "London", "Frankfurt", "Berlin"),
    stackID = c("capital city", "city", "capital city", 1, 2, 2, "A", "B", "A")
)

# Retrieve unique sorters without printing
printSorters(CMData, verbose = FALSE)

# Retrieve and print unique sorters to console
printSorters(CMData)

}
